coda_minigrid_project_coda-fixed/requirements.txt
gymnasium>=0.28
minigrid>=2.3.0
stable-baselines3>=2.3.0
sb3-contrib>=2.3.0
numpy>=1.24
pyyaml>=6.0
mpmath>=1.3.0
